requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.8.3
charset-normalizer==3.4.3
deprecation==2.1.0
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
frozenlist==1.8.0
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.11.0
lance-namespace==0.0.6
lance-namespace-urllib3-client==0.0.17
lancedb==0.25.1
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.2.6
openai==2.1.0
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
overrides==7.7.0
packaging==25.0
propcache==0.4.0
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pylance==0.38.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.5
pytimeparse==1.1.8
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
xai-sdk==1.2.0
yarl==1.22.0
zipp==3.23.0
